added data to NewEventForm.js, made modifications and fixed errors and finally passed all tests
